챌린지별 유저 풀이여부 작성

+++ SSSP/api/schemas/schema_score.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 class ScoreResponse(BaseModel):
     username: str
-    total_score: int
+    total_score: float
 
     class Config:
         from_attributes=True
